refactor(e2e): rename folder
