+++ __e2e__/user.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
 
-import { loginAsGuest } from '@tests/utils';
+import { loginAsGuest } from '@__e2e__/utils';
 
 test.beforeEach(async ({ page }) => {
   await loginAsGuest(page);
